pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fayda-auth"
version = "0.1.0"
description = "A Python library for Fayda Esignet OAuth authentication"
readme = "README.md"
authors = [
    { name = "Awura Computing PLC", email = "contact@awuraplc.org" },
]
license = { text = "MIT" }
requires-python = ">=3.8"
keywords = ["oauth", "fayda", "authentication", "esignet"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "redis>=4.0.0",
    "python-dotenv>=0.19.0",
    "pyjwt>=2.4.0",
    "cryptography>=36.0.0",
    "requests>=2.26.0",
    "jwcrypto>=1.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
]

# [project.urls]
# "Homepage" = ""
# "Repository" = ""
# "Documentation" = ""  # Optional

[tool.setuptools]
packages = ["fayda_auth"]
include-package-data = true

[tool.setuptools.package-data]
"fayda_auth" = ["*.py"]